+++ internal/repository/auth/auth_test.go
@@ -15,7 +15,33 @@ import (
 	"path"
 	"runtime"
 	"testing"
+
+	_ "github.com/lib/pq"
 )
+
+var mockDB *sql.DB
+var mockRepo *RepositoryAuth
+
+func TestMain(m *testing.M) {
+	if err := changeDirToRoot(); err != nil {
+		log.Fatal(err)
+	}
+
+	cfg, err := loadConfig()
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	db, err := newDBConnection(cfg)
+	mockDB = db
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	mockRepo = NewRepoAuth(mockDB)
+
+	os.Exit(m.Run())
+}
 
 func TestRepositoryAuth_CreateUser(t *testing.T) {
 	ctx := context.Background()
